Refactor map component to reduce code duplication
